+++ internal/config/core.go
@@ -1,6 +1,10 @@
 package config
 
+import (
+	"github.com/google/uuid"
+)
+
 type Core struct {
-	CoreURL      string `env:"CORE_URL" envDefault:""`
-	SubscriberID string `env:"CORE_SUBSCRIBER_ID" envDefault:""`
+	CoreURL      string    `env:"CORE_URL" envDefault:""`
+	SubscriberID uuid.UUID `env:"CORE_SUBSCRIBER_ID" envDefault:""`
 }