+++ internal/subscription/cache_test.go
@@ -3,6 +3,7 @@ package subscription
 import (
 	"testing"
 
+	"github.com/google/uuid"
 	"github.com/stretchr/testify/require"
 )
 
@@ -14,8 +15,13 @@ func TestUnitNewCache(t *testing.T) {
 
 func TestUnitAddAndGetItems(t *testing.T) {
 	c := NewCache()
-	c.AddItems("key-1", "val-1", "val-2")
-	c.AddItems("key-2", "val-0")
+
+	id1 := uuid.New()
+	id2 := uuid.New()
+	id3 := uuid.New()
+
+	c.AddItems("key-1", id1, id2)
+	c.AddItems("key-2", id3)
 	c.AddItems("key-3")
 
 	t.Run("length is valid", func(t *testing.T) {
@@ -31,15 +37,15 @@ func TestUnitAddAndGetItems(t *testing.T) {
 	t.Run("contains correct values", func(t *testing.T) {
 		items, ok := c.GetItems("key-1")
 		require.True(t, ok)
-		require.Contains(t, items, "val-1")
-		require.Contains(t, items, "val-2")
+		require.Contains(t, items, id1)
+		require.Contains(t, items, id2)
 	})
 }
 
 func TestUnitRemoveKey(t *testing.T) {
 	c := NewCache()
 
-	c.AddItems("key-1", "val-1", "val-2")
+	c.AddItems("key-1", uuid.New(), uuid.New())
 	c.RemoveKey("key-1")
 
 	items, ok := c.GetItems("key-1")
@@ -50,11 +56,14 @@ func TestUnitRemoveKey(t *testing.T) {
 func TestUnitRemoveItem(t *testing.T) {
 	c := NewCache()
 
-	c.AddItems("key-1", "val-1", "val-2")
-	c.RemoveItem("key-1", "val-2")
+	id1 := uuid.New()
+	id2 := uuid.New()
+
+	c.AddItems("key-1", id1, id2)
+	c.RemoveItem("key-1", id1)
 
 	items, ok := c.GetItems("key-1")
 	require.True(t, ok)
 	require.Len(t, items, 1)
-	require.Equal(t, items, []string{"val-1"})
+	require.Equal(t, items, []uuid.UUID{id2})
 }