+++ resources/schema.sql
@@ -1,34 +1,29 @@
 create table users
 (
-    id          text not null
-        primary key,
+    id          uuid not null primary key,
     created_at  timestamp with time zone,
     updated_at  timestamp with time zone,
     deleted_at  timestamp with time zone,
     device_uuid text
 );
 
-create index if not exists idx_users_deleted_at
-    on users (deleted_at);
+create index if not exists idx_users_deleted_at on users (deleted_at);
 
 create table user_subscriptions
 (
-    id          text not null
-        primary key,
+    id         uuid not null primary key,
     created_at timestamp with time zone,
     updated_at timestamp with time zone,
     deleted_at timestamp with time zone,
-    user_id    text,
-    dao_id     text
+    user_id    uuid,
+    dao_id     uuid
 );
 
-create index if not exists idx_subscriptions_deleted_at
-    on user_subscriptions (deleted_at);
+create index if not exists idx_subscriptions_deleted_at on user_subscriptions (deleted_at);
 
 create table global_subscriptions
 (
-    id            bigserial
-        primary key,
+    id            uuid primary key,
     created_at    timestamp with time zone,
     updated_at    timestamp with time zone,
     deleted_at    timestamp with time zone,
@@ -36,5 +31,4 @@ create table global_subscriptions
     dao_id        text
 );
 
-create index if not exists idx_global_subscriptions_deleted_at
-    on global_subscriptions (deleted_at);
+create index if not exists idx_global_subscriptions_deleted_at on global_subscriptions (deleted_at);